+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVM1HQVSNR15KAD030X8MNP3"
+        "01JVM4NY4MHTRCMBEBW0HMFPEJ"
     )
     r"""The id of the resource"""