+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -223,5 +223,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-19T10:30:59.370Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-19T11:25:42.827Z")
     r"""The date and time the resource was last updated"""